2003-08-25  Havoc Pennington  <hp@pobox.com>
        Just noticed that dbus_message_test is hosed, I wonder when I
	broke that. I thought make check was passing earlier...

	* dbus/dbus-object-tree.c: add new "object tree" to match DCOP
	container tree, will replace most of dbus-object-registry

	* dbus/dbus-string.c (_dbus_string_append_printf_valist): fix C99
	screwup

+++ dbus/dbus-message-builder.c
@@ -411,7 +411,7 @@ _dbus_message_data_load (DBusString       *dest,
           DBusString name;
           int message_type;
 
-          if (_dbus_string_get_length (&line) < strlen ("VALID_HEADER "))
+          if (_dbus_string_get_length (&line) < (int) strlen ("VALID_HEADER "))
             {
               _dbus_warn ("no args to VALID_HEADER\n");
               goto parse_failed;
